issue: 1404279 Fix TCP sockets of secondary IPs are not offloaded
As for UDP, from this release, ring can be share between several IPs.
hence, tcp-flow-map should be based also on dst_ip.

+++ src/vma/dev/rfs.h
@@ -203,9 +203,9 @@ typedef std::vector<attach_flow_data_t*> attach_flow_data_vector_t;
 class rfs_rule_filter
 {
 public:
-	rfs_rule_filter(rule_filter_map_t& map, uint32_t key, flow_tuple& flow_tuple) : m_map(map), m_key(key), m_flow_tuple(flow_tuple) {}
+	rfs_rule_filter(rule_filter_map_t& map, uint64_t key, flow_tuple& flow_tuple) : m_map(map), m_key(key), m_flow_tuple(flow_tuple) {}
 	rule_filter_map_t& m_map;
-	uint32_t m_key;
+	uint64_t m_key;
 	flow_tuple m_flow_tuple;
 };
